+++ src/common/gitProvider/index.ts
@@ -92,6 +92,30 @@ export abstract class GitProvider {
       return null;
     }
   }
+
+  static async getCurrentBranchUrl(): Promise<string> {
+    const gitProvider = GitProvider.getInstance();
+    if (gitProvider == null) {
+      return null;
+    }
+    return gitProvider.getCurrentBranchUrl();
+  }
+
+  static async getJobUrl(): Promise<string> {
+    const gitProvider = GitProvider.getInstance();
+    if (gitProvider == null) {
+      return null;
+    }
+    return gitProvider.getCurrentJobUrl();
+  }
+
+  static async getPullRequestInfo(): Promise<any> {
+    const gitProvider = GitProvider.getInstance();
+    if (gitProvider == null) {
+      return null;
+    }
+    return gitProvider.getPullRequestInfo();
+  }
 }
 
 export declare type PullRequestMessageRequest = {