+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
-- Native BitBucket compliance for PR deployment checks and deployments to major orgs after merge
+- sfdx-hardis & Slack Integration
+  - Easy configuration
+  - Deployment notifications to a common channel, and also to git branch dedicated channel
 
-- Added new option --testlevel RunRepositoryTests which will dynamically detect all GIT repository test classes and runs the deployment with found tests. This will speed up the validation/deployment on cases where GIT repository module contains subset of all tests found in the org
-- Added --runtests support in order to pass certain APEX test classes when --testlevel RunSpecifiedTests is used
+- **hardis:project:deploy:dx** enhancements:
+  - Added new option --testlevel RunRepositoryTests which will dynamically detect all GIT repository test classes and runs the deployment with found tests. This will speed up the validation/deployment on cases where GIT repository module contains subset of all tests found in the org
+  - Added --runtests support in order to pass certain APEX test classes when --testlevel RunSpecifiedTests is used
+- Native BitBucket compliance for PR deployment checks and deployments to major orgs after merge
 
 ## [4.5.1] 2023-09-11